+++ Mario/AttackablePower.cpp
@@ -88,7 +88,7 @@ void AttackablePower::Render()
 
 		if (IsAttacking() && m->GetMovingState() != MovingStates::CROUCH) {
 			selectedAnimation = animations["Attack"];
-			selectedAnimation->GetTransform()->Scale = Vec2(m->GetFacing(), 1);
+			selectedAnimation->GetTransform()->Scale = Vec2((float)m->GetFacing(), 1);
 			selectedAnimation->GetTransform()->Position = m->GetPosition() - cam;
 			selectedAnimation->Render(attackTimer.Elapsed(), MARIO_ATTACK_TIME);
 			return;
@@ -98,8 +98,8 @@ void AttackablePower::Render()
 			selectedAnimation = animations["Crouch"];
 		}		
 
-		selectedAnimation->SetPlayScale(max(0.4f, min(abs(m->GetVelocity().x) / MARIO_WALK_SPEED, 4)) * 1.5);
-		selectedAnimation->GetTransform()->Scale = Vec2(m->GetFacing(), 1);
+		selectedAnimation->SetPlayScale(max(0.4f, min(abs(m->GetVelocity().x) / MARIO_WALK_SPEED, 4)) * 1.5f);
+		selectedAnimation->GetTransform()->Scale = Vec2((float)m->GetFacing(), 1);
 		selectedAnimation->GetTransform()->Position = m->GetPosition() - cam;
 		selectedAnimation->Render();
 	}
